+++ src/test/java/study/calculator/StringAddCalculatorTest.java
@@ -1,9 +1,10 @@
-package study;
+package study.calculator;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.*;
+import study.calculator.StringAddCalculator;
 
 import java.util.stream.Stream;
