+++ src/test/java/study/learningtest/StringTest.java
@@ -1,4 +1,4 @@
-package study;
+package study.learningtest;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
